+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `rotate(angle: Scalar) -> Rotor`
+- `scale(factor: Scalar) -> Rotor`
+
 ## 0.3.0 - 2024-09-24
 
 ### Added